feat(phase2-3): api client split, decimal/timezone libs, statistics extraction, dashboard settings split, infoRepository split, docs sync
Phase 2 — Type safety:
- OpenAPI spec (openapi.yaml 1281 lines) + generated types (types/generated.ts)
- lib/decimal.ts + lib/timezone.ts: safe monetary arithmetic and date handling
- api.ts → lib/api/ barrel with 19 domain modules (aggregations, ai, categories,
  charts, electron, helpers, imports, info, market, planned, portfolio,
  recipients, settings, splits, transactions); api.ts is now a 191-line barrel
- ApiClientError + unwrapEnvelope<T> in lib/api/client.ts

Phase 3 — Perf hotspots:
- infoRepository.js split: 1443-line monolith → barrel + 4 domain modules
  (infoRepositoryNetWorth, infoRepositoryPlanned, infoRepositoryRecipients,
  infoRepositoryStatistics)
- StatisticsPage: 960-line page → 232-line orchestrator + 10 sub-components
  in components/statistics/ (CategoryPieChart, CategoryPivotTable,
  CategoryTrendChart, ChartCard, MonthlyChart, NetTrendChart,
  SavedChartsSection, SummaryCards, TopRecipientsChart, YearlySummaryTable)
- DashboardSettingsDialog: 1327-line file → 183-line orchestrator + 5 tab
  sub-components in components/settings/tabs/
- useChartCurrencyFormatter hook extracted from statistics consumers
- AIChatSettingsSection component extracted

Docs:
- ADR-031 (OpenAPI type generation), updated ADR index
- API docs for all 15+ route files
- Architecture, components, features, performance, security, testing docs synced
- api-endpoint-matrix.md, code-patterns.md, frontend-api-client.md updated

+++ apps/frontend/src/components/portfolio/AddInvestmentFromMarketDialog.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { parseDecimal } from '@/lib/decimal';
 import { useLanguage } from '@/contexts/LanguageContext';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
@@ -111,7 +112,7 @@ export function AddInvestmentFromMarketDialog({ quote, existingInvestment }: Pro
         symbol: newInvestmentForm.symbol.trim(),
         asset_class: assetClass,
         currency: newInvestmentForm.currency,
-        current_price: parseFloat(newInvestmentForm.currentPrice),
+        current_price: parseDecimal(newInvestmentForm.currentPrice),
         notes: newInvestmentForm.notes.trim() || undefined,
         price_provider: 'yahoo',
         price_provider_id: quote.symbol,
@@ -125,15 +126,15 @@ export function AddInvestmentFromMarketDialog({ quote, existingInvestment }: Pro
   };
 
   const computedAmount = transactionForm.units && transactionForm.pricePerUnit
-    ? (parseFloat(transactionForm.units) * parseFloat(transactionForm.pricePerUnit)).toFixed(2)
+    ? (parseDecimal(transactionForm.units) * parseDecimal(transactionForm.pricePerUnit)).toFixed(2)
     : '';
 
   const handleAddTransaction = async (e: React.FormEvent) => {
     e.preventDefault();
     if (!existingInvestment) return;
     
-    const amount = parseFloat(transactionForm.amount || computedAmount);
-    if (!amount || isNaN(amount)) { 
+    const amount = parseDecimal(transactionForm.amount || computedAmount);
+    if (!amount || isNaN(amount)) {
       toast.error(t('addPortTxn.error.amountRequired'));
       return; 
     }
@@ -144,10 +145,10 @@ export function AddInvestmentFromMarketDialog({ quote, existingInvestment }: Pro
         type: transactionForm.type,
         date: transactionForm.date,
         amount,
-        units: transactionForm.units ? parseFloat(transactionForm.units) : undefined,
-        price_per_unit: transactionForm.pricePerUnit ? parseFloat(transactionForm.pricePerUnit) : undefined,
-        fees: transactionForm.fees ? parseFloat(transactionForm.fees) : undefined,
-        taxes: transactionForm.taxes ? parseFloat(transactionForm.taxes) : undefined,
+        units: transactionForm.units ? parseDecimal(transactionForm.units) : undefined,
+        price_per_unit: transactionForm.pricePerUnit ? parseDecimal(transactionForm.pricePerUnit) : undefined,
+        fees: transactionForm.fees ? parseDecimal(transactionForm.fees) : undefined,
+        taxes: transactionForm.taxes ? parseDecimal(transactionForm.taxes) : undefined,
         currency: existingInvestment.currency,
         note: transactionForm.note.trim() || undefined,
         is_recurring: transactionForm.isRecurring,
